pyproject.toml
[project]
name = "pedidos-verduleria-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.10",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.5.3",
    "pillow>=12.1.0",
]
